+++ src/renderer/src/features/chat/components/chat-sidebar.tsx
@@ -16,19 +16,12 @@ import {
 	alpha,
 } from "@mui/material";
 import { styled } from "@mui/material/styles";
-import { useExportAgent } from "@shared/hooks/use-agent-mutations";
-import { useAgents } from "@shared/hooks/use-agents";
-import { useClearAgentConversation } from "@shared/hooks/use-clear-agent-conversation";
-import { usePaginationParams } from "@shared/hooks/use-pagination-params";
+import { useExportAgent, useAgents, useClearAgentConversation, usePaginationParams } from "@shared/hooks";
 import {
 	formatMessageDateTime,
 	getFullDateTime,
 } from "@shared/utils/date-utils";
-import { AgentOptionsMenu } from "@shared/components/common/agent-options-menu";
-import { CompactPagination } from "@shared/components/common/compact-pagination";
-import { CreateAgentDialog } from "@shared/components/common/create-agent-dialog";
-import { ImportAgentDialog } from "@shared/components/common/import-agent-dialog";
-import { SidebarHeader } from "@shared/components/common/sidebar-header";
+import { AgentOptionsMenu, CompactPagination, CreateAgentDialog, ImportAgentDialog, SidebarHeader } from "@shared/components/common";
 import type { ChangeEvent, FC } from "react";
 import { useCallback, useState } from "react";
