+++ src/renderer/src/shared/components/common/index.ts
@@ -1,8 +1,25 @@
-export {
-	UpdateContainer,
-	UpdateActions,
-	ProgressContainer,
-	UpdateNotification,
-} from "./update-notification";
+export { AgentOptionsMenu } from "./agent-options-menu";
+export { AutocompleteField } from "./autocomplete-field";
+export { BaseDialog } from "./base-dialog";
+export { ChatLayout } from "./chat-layout";
+export { CheckForUpdatesButton } from "./check-for-updates-button";
+export { CompactPagination } from "./compact-pagination";
+export { ConfirmationModal } from "./confirmation-modal";
 export { ConnectivityBanner } from "./connectivity-banner";
+export { CreateAgentDialog } from "./create-agent-dialog";
+export { EditableField } from "./editable-field";
+export { ErrorBoundary } from "./error-boundary";
+export { ImportAgentDialog } from "./import-agent-dialog";
+export { ModelsInitializer } from "./models-initializer";
+export { PageHeader } from "./page-header";
+export { ResizableDivider } from "./resizable-divider";
+export { SidebarHeader } from "./sidebar-header";
+export { SliderSetting } from "./slider-setting";
 export { ThemedToastContainer } from "./themed-toast-container";
+export { ToggleSetting } from "./toggle-setting";
+export {
+  UpdateContainer,
+  UpdateActions,
+  ProgressContainer,
+  UpdateNotification,
+} from "./update-notification";
