+++ src/renderer/src/shared/components/hosting/model-select.tsx
@@ -17,7 +17,7 @@ import {
 	createFilterOptions,
 	styled,
 } from "@mui/material";
-import { useModels } from "@shared/hooks/use-models";
+import { useModels } from "@shared/hooks";
 import type { FC, SyntheticEvent } from "react";
 import { useEffect, useMemo, useRef, useState } from "react";
 import ReactMarkdown from "react-markdown";
